requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==3.2.0
bleak==1.1.0
certifi==2025.4.26
cffi==2.0.0
click==7.1.2
clr_loader==0.2.7.post0
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.124.0
Flask==1.1.2
gunicorn==20.1.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.18.1
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==2.11.3
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.7
motor==3.7.1
nest-asyncio==1.6.0
numpy==1.20.2
packaging==25.0
pandas==1.2.4
parso==0.8.4
passlib==1.7.4
Pillow==8.2.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pybluez2==0.46
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pythonnet==3.0.1
pytz==2021.1
pywin32==311
pyzmq==27.0.0
rsa==4.9.1
schedule==1.2.2
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.49.3
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
uvicorn==0.38.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==1.0.1
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
wsproto==1.2.0
zipp==3.23.0
